tests integ: Introduce the bondlib module
Extract the bond creation function to a dedicated module, enabling
it to be re-used in other tests.
